web: move api handlers to package 'api'
