+++ internal/web/api/utils.go
@@ -1,0 +1,12 @@
+package api
+
+import "strings"
+
+// trimSpacePointer is like 'strings.TrimPointer' but for pointers
+func trimSpacePointer(s *string) *string {
+	if s == nil {
+		return nil
+	}
+	trimmed := strings.TrimSpace(*s)
+	return &trimmed
+}
