+++ internal/web/routes.go
@@ -5,6 +5,8 @@ import (
 	"net/http/pprof"
 
 	"github.com/gorilla/mux"
+
+	"github.com/ShoshinNikita/budget-manager/internal/web/api"
 )
 
 type route struct {
@@ -14,6 +16,8 @@ type route struct {
 }
 
 func (s Server) addRoutes(router *mux.Router) {
+	apiHandlers := api.NewHandlers(s.db, s.log)
+
 	routes := []route{
 		// Pages
 		{methods: "GET", path: "/overview", handler: s.overviewPage},
@@ -24,29 +28,29 @@ func (s Server) addRoutes(router *mux.Router) {
 		{methods: "GET", path: "/", handler: s.indexHandler},
 
 		// API
-		{methods: "GET", path: "/api/months/id", handler: s.GetMonthByID},
-		{methods: "GET", path: "/api/months/date", handler: s.GetMonthByDate},
-		{methods: "GET", path: "/api/days/id", handler: s.GetDayByID},
-		{methods: "GET", path: "/api/days/date", handler: s.GetDayByDate},
+		{methods: "GET", path: "/api/months/id", handler: apiHandlers.GetMonthByID},
+		{methods: "GET", path: "/api/months/date", handler: apiHandlers.GetMonthByDate},
+		{methods: "GET", path: "/api/days/id", handler: apiHandlers.GetDayByID},
+		{methods: "GET", path: "/api/days/date", handler: apiHandlers.GetDayByDate},
 		// Income
-		{methods: "POST", path: "/api/incomes", handler: s.AddIncome},
-		{methods: "PUT", path: "/api/incomes", handler: s.EditIncome},
-		{methods: "DELETE", path: "/api/incomes", handler: s.RemoveIncome},
+		{methods: "POST", path: "/api/incomes", handler: apiHandlers.AddIncome},
+		{methods: "PUT", path: "/api/incomes", handler: apiHandlers.EditIncome},
+		{methods: "DELETE", path: "/api/incomes", handler: apiHandlers.RemoveIncome},
 		// Monthly Payment
-		{methods: "POST", path: "/api/monthly-payments", handler: s.AddMonthlyPayment},
-		{methods: "PUT", path: "/api/monthly-payments", handler: s.EditMonthlyPayment},
-		{methods: "DELETE", path: "/api/monthly-payments", handler: s.RemoveMonthlyPayment},
+		{methods: "POST", path: "/api/monthly-payments", handler: apiHandlers.AddMonthlyPayment},
+		{methods: "PUT", path: "/api/monthly-payments", handler: apiHandlers.EditMonthlyPayment},
+		{methods: "DELETE", path: "/api/monthly-payments", handler: apiHandlers.RemoveMonthlyPayment},
 		// Spend
-		{methods: "POST", path: "/api/spends", handler: s.AddSpend},
-		{methods: "PUT", path: "/api/spends", handler: s.EditSpend},
-		{methods: "DELETE", path: "/api/spends", handler: s.RemoveSpend},
+		{methods: "POST", path: "/api/spends", handler: apiHandlers.AddSpend},
+		{methods: "PUT", path: "/api/spends", handler: apiHandlers.EditSpend},
+		{methods: "DELETE", path: "/api/spends", handler: apiHandlers.RemoveSpend},
 		// Spend Type
-		{methods: "GET", path: "/api/spend-types", handler: s.GetSpendTypes},
-		{methods: "POST", path: "/api/spend-types", handler: s.AddSpendType},
-		{methods: "PUT", path: "/api/spend-types", handler: s.EditSpendType},
-		{methods: "DELETE", path: "/api/spend-types", handler: s.RemoveSpendType},
+		{methods: "GET", path: "/api/spend-types", handler: apiHandlers.GetSpendTypes},
+		{methods: "POST", path: "/api/spend-types", handler: apiHandlers.AddSpendType},
+		{methods: "PUT", path: "/api/spend-types", handler: apiHandlers.EditSpendType},
+		{methods: "DELETE", path: "/api/spend-types", handler: apiHandlers.RemoveSpendType},
 		// Other
-		{methods: "GET", path: "/api/search/spends", handler: s.SearchSpends},
+		{methods: "GET", path: "/api/search/spends", handler: apiHandlers.SearchSpends},
 	}
 	for _, r := range routes {
 		router.Methods(r.methods).Path(r.path).HandlerFunc(r.handler)
